+++ src/main/java/pl/storex/storex/LoadDatabase.java
@@ -2,15 +2,10 @@ package pl.storex.storex;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.event.EventListener;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import pl.storex.storex.user.User;
-import pl.storex.storex.user.UserDTO;
-import pl.storex.storex.user.UserRepository;
-import pl.storex.storex.user.UserService;
+import pl.storex.storex.user.model.UserDTO;
+import pl.storex.storex.user.service.UserService;
 
 public class LoadDatabase {
     private static final Logger logger = LoggerFactory.getLogger(LoadDatabase.class);
