+++ src/main/java/pl/storex/storex/user/model/Role.java
@@ -1,4 +1,4 @@
-package pl.storex.storex.user;
+package pl.storex.storex.user.model;
 
 public enum Role {
     USER, ADMIN
