+++ src/main/java/pl/storex/storex/user/model/UserDTO.java
@@ -1,4 +1,4 @@
-package pl.storex.storex.user;
+package pl.storex.storex.user.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
